kernel/sysctl.c: add flags to support min/max range clamping
Patch series "pc: Clamp *mni to the real IPCMNI limit", v4.

The sysctl parameters msgmni, shmmni and semmni have an inherent limit of
IPC_MNI (32k).  However, users may not be aware of that because they can
write a value much higher than that without getting any error or
notification.  Reading the parameters back will show the newly written
values which are not real.

Enforcing the limit by failing sysctl parameter write, however, can break
existing user applications.  To address this delemma, a new flags field is
introduced into the ctl_table.  The value CTL_FLAGS_CLAMP_RANGE can be
added to any ctl_table entries to enable a looser range clamping without
returning any error.  For example,

  .flags = CTL_FLAGS_CLAMP_RANGE,

This flags value are now used for the range checking of shmmni, msgmni and
semmni without breaking existing applications.  If any out of range value
is written to those sysctl parameters, the following warning will be
printed instead.

  sysctl: "shmmni" was set out of range [0, 32768], clamped to 32768.

Reading the values back will show 32768 instead of some fake values.

This patch (of 6):

When minimum/maximum values are specified for a sysctl parameter in the
ctl_table structure with proc_dointvec_minmax() handler, update to that
parameter will fail with error if the given value is outside of the
required range.

There are use cases where it may be better to clamp the value of the
sysctl parameter to the given range without failing the update, especially
if the users are not aware of the actual range limits.  Reading the value
back after the update will now be a good practice to see if the provided
value exceeds the range limits.

To provide this less restrictive form of range checking, a new flags field
is added to the ctl_table structure.

When the CTL_FLAGS_CLAMP_RANGE flag is set in the ctl_table entry, any
update from the userspace will be clamped to the given range without error
if either the proc_dointvec_minmax() or the proc_douintvec_minmax()
handlers is used.

Link: http://lkml.kernel.org/r/1520885744-1546-2-git-send-email-longman@redhat.com
Signed-off-by: Waiman Long <longman@redhat.com>
Reviewed-by: Andrew Morton <akpm@linux-foundation.org>
Cc: "Luis R. Rodriguez" <mcgrof@kernel.org>
Cc: Kees Cook <keescook@chromium.org>
Cc: Al Viro <viro@zeniv.linux.org.uk>
Cc: Matthew Wilcox <willy@infradead.org>
Cc: "Eric W. Biederman" <ebiederm@xmission.com>
Cc: Alexey Dobriyan <adobriyan@gmail.com>
Cc: Davidlohr Bueso <dave@stgolabs.net>
Cc: Manfred Spraul <manfred@colorfullife.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Stephen Rothwell <sfr@canb.auug.org.au>