arm: dts: armada-xp-netgear-rn2120: use SPDX-License-Identifier
Follow the recent trend for the license description, and also fix the
wrongly stated X11 to MIT.

As already pointed on the DT ML, the X11 license text [1] is explicitly
for the X Consortium and has a couple of extra clauses. The MIT
license text [2] is actually what the current DT files claim.

[1] https://spdx.org/licenses/X11.html
[2] https://spdx.org/licenses/MIT.html

Cc: Arnaud Ebalard <arno@natisbad.org>
Signed-off-by: Gregory CLEMENT <gregory.clement@bootlin.com>