arm: dts: armada-38x: use SPDX-License-Identifier for Armada 38x SoCs
Follow the recent trend for the license description, and also fix the
wrongly stated X11 to MIT.

As already pointed on the DT ML, the X11 license text [1] is explicitly
for the X Consortium and has a couple of extra clauses. The MIT
license text [2] is actually what the current DT files claim.

The licence of the arch/arm/boot/dts/armada-388.dtsi file was also
modified to GPLv2+ instead of GPLv2. Actually the GPLv2 was a mistake
that I can fix being the author of this file.

[1] https://spdx.org/licenses/X11.html
[2] https://spdx.org/licenses/MIT.html

Signed-off-by: Gregory CLEMENT <gregory.clement@bootlin.com>