arm: dts: armada-385-turris-omnia: use SPDX-License-Identifier
Follow the recent trend for the license description, and also fix the
wrongly stated X11 to MIT.

As already pointed on the DT ML, the X11 license text [1] is explicitly
for the X Consortium and has a couple of extra clauses. The MIT
license text [2] is actually what the current DT files claim.

[1] https://spdx.org/licenses/X11.html
[2] https://spdx.org/licenses/MIT.html

Cc: Uwe Kleine-König <uwe@kleine-koenig.org>
Signed-off-by: Gregory CLEMENT <gregory.clement@bootlin.com>