of/fdt: rework early_init_dt_scan_chosen
The FDT scanning functions iterate over nodes with an iterator function
callback, but in many cases we just need to find the node for an
absolute path and libfdt provides such a function. Split out the node
searching and property parsing code to 2 functions and use
fdt_path_offset() to get the chosen node.

This also has the side effect of only checking for "/chosen@0" on
powerpc which should be the only systems that possibly have that node.

Cc: Benjamin Herrenschmidt <benh@kernel.crashing.org>
Cc: Paul Mackerras <paulus@samba.org>
Cc: Michael Ellerman <mpe@ellerman.id.au>
Cc: Frank Rowand <frowand.list@gmail.com>
Cc: linuxppc-dev@lists.ozlabs.org
Signed-off-by: Rob Herring <robh@kernel.org>