xfs: do not log swapext extent owner changes for deleted
Today if we run fsr and crash, log replay can fail because
the recovery code tries to instantiate the donor inode from
disk to replay the swapext, but it's been deleted and we get
verifier failures when we try to read it off disk with
i_mode == 0.

Strip the extent owner changes out of the logged fields when
we're freeing the inode to avoid this.

Signed-off-by: Eric Sandeen <sandeen@redhat.com>
Reviewed-by: Darrick J. Wong <darrick.wong@oracle.com>
Signed-off-by: Darrick J. Wong <darrick.wong@oracle.com>