dm verity: aggregate crypto API calls
Current implementation makes multiple crypto API calls per block, which
forces the underlying crypto tfm implementation to "stop & go".  This
leads to under-utilization of of the crypto HW engines.

Fix this by unifying the calls to crypto init/update/final into a digest
call with a single sg which contains multiple buffers.  This is also
needed as an enabler for the next patch in the series.

Signed-off-by: Yael Chemla <yael.chemla@foss.arm.com>
Reviewed-by: Gilad Ben-Yossef <gilad@benyossef.com>
Signed-off-by: Mike Snitzer <snitzer@redhat.com>