of/fdt: Respect CONFIG_CMDLINE{,_EXTENED,_FORCE) with no chosen node
The current behavior of early_init_dt_scan_chosen filling in
boot_command_line varies if chosen node is not present compared to if
just bootargs is not present. If chosen node is not present,
boot_command_line is not touched and we don't honor CONFIG_CMDLINE*
options. Change this to be consistent with just bootargs being not
present and fill in boot_command_line based on the config options.

Reported-by: Palmer Dabbelt <palmer@sifive.com>
Cc: Trung Tran <trung.tran@ettus.com>
Cc: Moritz Fischer <mdf@kernel.org>
Cc: Michael Clark <mjc@sifive.com>
Cc: Frank Rowand <frowand.list@gmail.com>
Signed-off-by: Rob Herring <robh@kernel.org>