MIPS: vmlinuz: Use generic ashldi3 and build directly
In preparation for removing some of the MIPS compiler intrinsics from
arch/mips/lib, first update the build of vmlinuz to use the generic
ashldi3 from lib.

Both ashldi3 and bswapsi objects need to be built with different CFLAGS
for inclusion to vmlinuz rather than simply including the object built
for the main kernel image. But the current copy of the source C file to
arch/mips/boot/compressed can be avoided by simply calling cmd,cc_o_c to
build the object from the source directly. This also removes the need
for the .gitignore file to ignore the copied files, and the extra-y rule
to clean them.

Signed-off-by: Matt Redfearn <matt.redfearn@mips.com>
Cc: Ralf Baechle <ralf@linux-mips.org>
Cc: Alban Bedel <albeu@free.fr>
Cc: Antony Pavlov <antonynpavlov@gmail.com>
Cc: Palmer Dabbelt <palmer@sifive.com>
Cc: Bartosz Golaszewski <brgl@bgdev.pl>
Cc: linux-mips@linux-mips.org
Patchwork: https://patchwork.linux-mips.org/patch/18999/
Signed-off-by: James Hogan <jhogan@kernel.org>