treewide: remove false-positive VLAs when using max()
As part of removing VLAs from the kernel[1], we want to build with -Wvla,
but it is overly pessimistic and only accepts constant expressions for
stack array sizes, instead of also constant values.  The max() macro
triggers the warning, so this refactors these uses of max() to use the new
const_max() instead.

[1] https://lkml.org/lkml/2018/3/7/621

Link: http://lkml.kernel.org/r/1521143266-31350-3-git-send-email-keescook@chromium.org
Signed-off-by: Kees Cook <keescook@chromium.org>
Reviewed-by: Nikolay Borisov <nborisov@suse.com> [btrfs]
Cc: Josh Poimboeuf <jpoimboe@redhat.com>
Cc: Rasmus Villemoes <linux@rasmusvillemoes.dk>
Cc: Randy Dunlap <rdunlap@infradead.org>
Cc: Miguel Ojeda <miguel.ojeda.sandonis@gmail.com>
Cc: Ingo Molnar <mingo@kernel.org>
Cc: David Laight <David.Laight@aculab.com>
Cc: Ian Abbott <abbotti@mev.co.uk>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Stephen Rothwell <sfr@canb.auug.org.au>