CHROMIUM: Bluetooth: increase LE auto connection timeout to 5 seconds
This changes the LE auto connection timeout from 2 seconds to 5 seconds
in order to improve the connection successful rate in a noisy
environment with the most common use case for Chromeb* - audio streaming
via Wi-Fi.
In the experiment, the LE connection attempts to a remote Android phone
(1000 times) were performed when the DUT was connected to a pair of
Bluetooth headphone and streamed audio via Wi-Fi. The observation is
that ~1.5% of the LE connection attempts took more than 2 seconds to
complete, and the max time is around 3.5 seconds. Therefore, 5 is
selected to be the timeout.

BUG=None
TEST=perform LE connection test for 1000 times with an Android phone while
streaming audio via Wi-Fi and through a pair of Bluetooth headphone

Change-Id: I545e74ac75db18cc4b7a9c05880c9765af77585c
Signed-off-by: Miao-chen Chou <mcchou@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1000097
Reviewed-by: Sukumar Ghorai <sukumar.ghorai@intel.com>
Reviewed-by: Dmitry Grinberg <dmitrygr@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1644677
Commit-Ready: Yoni Shavit <yshavit@chromium.org>
Tested-by: Yoni Shavit <yshavit@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>