gpio: mxc: Remove reliance on <linux/gpio.h>
This is a driver so we should only include <linux/gpio.h>.
However this driver was using gpio_get_value() to fetch the
current value of a GPIO used as IRQ line to determine trigger
direction, so we need a better way than looping over the
global GPIO numberspace.

Fix this by just calling the .get() function in the GPIO chip,
as we don't want to end up creating a consumer dependency
on ourselves.

Cc: Vladimir Zapolskiy <vz@mleia.com>
Cc: Fabio Estevam <fabio.estevam@nxp.com>
Cc: Shawn Guo <shawnguo@kernel.org>
Signed-off-by: Linus Walleij <linus.walleij@linaro.org>