nfsd4: support change_attr_type attribute
The change attribute is what is used by clients to revalidate their
caches.  Our server may use i_version or ctime for that purpose.  Those
choices behave slightly differently, and it may be useful to the client
to know which we're using.  This attribute tells the client that.  The
Linux client doesn't yet use this attribute yet, though.

XXX: use pynfs to fetch and display TIME_DELTA and CHANGE_ATTR_TYPE.

Signed-off-by: J. Bruce Fields <bfields@redhat.com>