mtd: rawnand: micron: Fix support for on-die ECC
It looks like the NAND_STATUS_FAIL bit is sticky after an ECC failure,
which leads all READ operations following the failing one to report
an ECC failure. Reset the chip to clear the NAND_STATUS_FAIL bit.

Note that this behavior is not document in the datasheet, but resetting
the chip is the only solution we found to fix the problem.

Fixes: 9748e1d87573 ("mtd: nand: add support for Micron on-die ECC")
Cc: <stable@vger.kernel.org>
Signed-off-by: Boris Brezillon <boris.brezillon@bootlin.com>
Cc: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
Cc: Bean Huo <beanhuo@micron.com>
Cc: Peter Pan <peterpandong@micron.com>
Reviewed-by: Miquel Raynal <miquel.raynal@bootlin.com>