htc_battery: Only load the driver on Walleye when built in
When the driver is built into the kernel image, it can register on
Taimen since there is no point where it rejects a non-Walleye battery;
rather, it just sets everything to a safe, default value. This will
ultimately conflict with lge_battery.

This allows a unified kernel image to properly work on both Taimen and
Walleye.

Change-Id: Ia894f61564df767b89c36485bfa3afbe8c1e644b
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>