btrfs: replace uuid_mutex with device list mutex in btrfs_free_extra_devids
btrfs_free_extra_devids() is called only in the mount context which
traverses through the fs_devices::devices and frees the orphan devices
in the given %fs_devices list, if any. As the search for the orphan
device is limited to fs_devices::devices so we don't need the global
uuid_mutex.

There can't be any mount-point based ioctl threads running yet as the
mount call is not yet finished. But there can be the btrfs-control
based scan ioctl running in parallel which calls device_list_add().

Here in the mount thread the fs_devices::opened is incremented way before
btrfs_free_extra_devids() is called and in the scan context the fs_devices
which are already opened neither be freed or alloc-able at
device_list_add().

But let's say you change the device-path and call the scan again, then
scan would update the new device path and this operation could race
against the btrfs_free_extra_devids() thread, which might be in the
process of freeing the same device. So synchronize it by using the
device_list_mutex.

This scenario is very rare, and practically the scan and mount are
anyway serialized by the usage so unless the race is instrumented its
very difficult to achieve.

Signed-off-by: Anand Jain <anand.jain@oracle.com>
Reviewed-by: David Sterba <dsterba@suse.com>
Signed-off-by: David Sterba <dsterba@suse.com>