PCI: Improve "partially hidden behind bridge" log message
There is a sanity checker in the end of pci_scan_child_bus_extend() that
tries to detect badly configured bridges.  For example given the below
topology:

  +-1b.0-[01-39]----00.0-[02-3a]--+-00.0-[03]----00.0
                                  +-01.0-[04-39]--
                                  \-02.0-[3a]----00.0

The sanity checker notices this and logs following messages:

  pci_bus 0000:3a: [bus 3a] partially hidden behind bridge 0000:02 [bus 02-39]
  pci_bus 0000:3a: [bus 3a] partially hidden behind bridge 0000:01 [bus 01-39]

This is not really helpful to users and the information above is not even
correct (0000:02 is a bus not bridge).  Make this a bit more understandable
by changing the sanity checker to log following message in place of the
above two messages:

  pci 0000:02:02.0: devices behind bridge are unusable because [bus 3a] cannot be assigned for them

While there update the comment on top of the sanity checker block to make
it clear that it is not only restricted to CardBus.

Suggested-by: Bjorn Helgaas <bhelgaas@google.com>
Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>
Signed-off-by: Bjorn Helgaas <helgaas@kernel.org>