autofs - make autofs4 Kconfig depend on AUTOFS_FS
Setting AUTOFS_FS and AUTOFS4_FS to y or m results in a compile failure.

It isn't possible to unset one if the other is set because it introduces a
circular dependency.

But adding "depends on AUTOFS_FS = n" will make autofs4 appear under the
autofs entry which should call attention to it.  It also results in
AUTOFS4_FS being removed from .config for any setting of AUTOFS_FS except
n.

Adding a "NOTE" to the AUTOFS4_FS help text will hopefully raise awareness
of the change to smooth the transition.

Link: http://lkml.kernel.org/r/152687649097.8263.7046086367407522029.stgit@pluto.themaw.net
Signed-off-by: Ian Kent <raven@themaw.net>
Reported-by: Randy Dunlap <rdunlap@infradead.org>
Tested-by: Randy Dunlap <rdunlap@infradead.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Stephen Rothwell <sfr@canb.auug.org.au>