msm: mdss: Fix redundant disp_err_detect IRQ enable on boot
The disp_err_detect IRQ is enabled by default on boot when it is first
requested, so when the panel is powered on during boot, the disp_err_detect
IRQ is redundantly enabled which causes a warning to be printed to dmesg.

To fix the warning, block attempts to enable/disable the disp_err_detect
IRQ redundantly.

Change-Id: I5f1a336e2166202dadc12e5884a40ecd4ce04ac4
Signed-off-by: Sultan Alsawaf <sultanxda@gmail.com>