CHROMIUM: tpm: cr50_i2c: Set hwrng quality.
This sets a quality score for the hwrng, causing it to be used
to populate the kernel's entropy pool.

The score was decided after discussion with the designer of the
rng, and is based on results from running the NIST SP800-90B
Entropy Assessment suite.

BUG=chromium:827682
TEST=build

Change-Id: I28a30f2be8466b59a0987c345d1ec7e93dbeb15d
Reviewed-on: https://chromium-review.googlesource.com/1082180
Commit-Ready: Louis Collard <louiscollard@chromium.org>
Tested-by: Louis Collard <louiscollard@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>