treewide: Fix code errors detected using GCC 8.1.0
Use the latest version of GCC to take advantage of improved static analysis.
These issues appeared as warnings from the compiler.

Included in this commit is a buffer overflow fix (in touch_sw49408_prd.h),
as well as fixes for other lethal issues.

Compiler warnings should not be taken lightly.

Change-Id: Ia3c7e32fb94891a9302c7f72dc71e97e5b37ce98
Signed-off-by: Sultan Alsawaf <sultanxda@gmail.com>