msm: mdss: Power on display asynchronously as early as possible
Currently, mdss is completely blocked while powering on the display, so if
the display takes a very long time to turn on, then mdss will be stuck
unable to do anything else in the meantime. This results in a long delay
between trying to wake the device and the display actually powering on and
rendering a frame.

In order to make the display turn on faster when waking the device from
sleep, start powering on the display as soon as the framebuffer unblank
event is received. This allows mdss to continue resuming while the display
takes its time powering on. A high-priority kthread is used here to ensure
the display powers on as fast as possible.

To make this work, the kickoff timeout needs to be relaxed because mdss
could attempt a kickoff quite a bit before the display is finished powering
on. Additionally, the kickoff in mdss_mdp_overlay_on() needs to be removed
because it occurs while the display power-on routines are still touching
with the dsi clocks, which results in several ugly error messages being
printed to dmesg. This kickoff is redundant since the display thread will
kickoff upon receiving the first frame after unblank anyway, so it's safe
to remove.

Change-Id: I7c9e64b98a35179f2164a2c903278e792e83c4d3
Signed-off-by: Sultan Alsawaf <sultanxda@gmail.com>