arm64: dts: taimen: Disable panel hardware error detection
Commit "arm64: dts: taimen: To avoid AOD blinking issues, changed some related commands"
causes spurious panel error interrupts to be fired off when the panel is
turning on, which causes the following error message:
[  405.446609] mdss_fb_report_panel_dead: Panel has gone bad, sending uevent - PANEL_ALIVE=0

Upon receiving the panel-death interrupt, panel recovery takes place which
causes the panel to remain powered off for an additional few seconds when
turning it on, or causes the panel to flash when turning it on.

Although the AOD command update is what's causing this issue, it's
important so it can't be easily removed. Instead of removing it, remove
the hardware-based panel error detection since it is no longer reliable.

Change-Id: I50dc5746b0fc03ca963385d20e5cb8c6ba132219
Signed-off-by: Sultan Alsawaf <sultanxda@gmail.com>