dt-bindings: soc: qcom: Fix default clock-freq for qcom,geni-i2c
In an early version of the I2C patch that was posted to the list the
default I2C frequency (if none was specified) was 400 kHz.  There was
debate on the list and we decided that it would be more consistent
with the rest of i2c if we defaulted to 100 kHz.  ...but we never
updated the bindings.  Let's fix this.

NOTE: since the i2c driver itself hasn't actually landed yet and the
SoC here is very new it seems terribly unlikely that anyone was
relying on the old 400 kHz number, so I'll assume this is an OK
"incompatible" device tree change.

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Bjorn Andersson <bjorn.andersson@linaro.org>
Acked-by: Wolfram Sang <wsa@the-dreams.de>
Signed-off-by: Andy Gross <andy.gross@linaro.org>