powerpc/pkeys: Preallocate execute-only key
The execute-only key is allocated dynamically. This is a problem. When
a thread implicitly creates an execute-only key, and resets UAMOR for
that key, the UAMOR value does not percolate to all the other threads.
Any other thread may ignorantly change the permissions on the key.
This can cause the key to be not execute-only for that thread.

Preallocate the execute-only key and ensure that no thread can change
the permission of the key, by resetting the corresponding bit in
UAMOR.

Fixes: 5586cf61e108 ("powerpc: introduce execute-only pkey")
Cc: stable@vger.kernel.org # v4.16+
Tested-by: Florian Weimer <fweimer@redhat.com>
Signed-off-by: Ram Pai <linuxram@us.ibm.com>
Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>