CHROMIUM: ASoC: DA7219: For Grunt save and restore state of register on S3
Reland

commit 6aa4fd739a6ff ("CHROMIUM: ASoC: DA7219: For Grunt save and restore state of register on S3")

The difference is that now we use component in codec driver.

On Grunt platform at S3 the power line to da7219 is cut-off. As a
result the codec is in reset state.
To provide a software workaround, the register state is saved at the
time of suspend and is restored on resume.

CQ-DEPEND=CL:1139472
BUG=b:109765944
TEST= play audio
      powerd_dbus_suspend
      resume
      audio continues to play

Signed-off-by: Akshu Agrawal <akshu.agrawal@amd.com>
Reviewed-on: https://chromium-review.googlesource.com/1109606
Reviewed-by: Simon Glass <sjg@chromium.org>
Signed-off-by: Cheng-Yi Chiang <cychiang@chromium.org>

Change-Id: I66b22d347ca7113a8274f6012b097743131701ad
Reviewed-on: https://chromium-review.googlesource.com/1139728
Commit-Ready: Cheng-Yi Chiang <cychiang@chromium.org>
Tested-by: Cheng-Yi Chiang <cychiang@chromium.org>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>