FROMLIST: z3fold: fix wrong handling of headless pages
During the processing of headless pages in z3fold_reclaim_page(),
there was a problem that the zhdr pointed to another page
or a page was already released in z3fold_free(). So, the wrong page
is encoded in headless, or test_bit does not work properly
in z3fold_reclaim_page(). This patch fixed these problems.

Signed-off-by: Jongseok Kim <ks77sj@gmail.com>
(am from https://patchwork.kernel.org/patch/10510583/)

BUG=chromium:822360
TEST=Boot with z3fold enabled

Change-Id: I2d7f60f415892272c41792e0271c8727e22129a3
Signed-off-by: Guenter Roeck <groeck@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1142465
Tested-by: Sonny Rao <sonnyrao@chromium.org>
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>